+++ lib/payloads/index.js
@@ -1,0 +1,9 @@
+const loginPayload = require('./loginPayload');
+const wssLoginPayload = require('./wssLoginPayload');
+const wssUpdatePayload = require('./wssUpdatePayload');
+
+module.exports = {
+  loginPayload,
+  wssLoginPayload,
+  wssUpdatePayload,
+};
